Move middleware into its own directory, add Wallet-API-Platform team to CODEOWNERS
