+++ app/scripts/lib/middleware/createLoggerMiddleware.js
