+++ app/scripts/lib/middleware/createMetaRPCHandler.js
