+++ app/scripts/lib/middleware/createOnboardingMiddleware.js
