+++ app/scripts/lib/middleware/createRPCMethodTrackingMiddleware.js
