+++ app/scripts/lib/middleware/createTabIdMiddleware.js
